Add a permits m2m field in Order model
Previously we were storing the latest order for the permit in the permit
model. This causes problems when a new order is created for the permit, we
lost of the track of the relationship between old orders and the permits.
Storing the permits in the Order model makes it easy to check the which
permits that each order is created for.

Refs: PV-375